Move S3DiscoveryMetrics out of metrics package
