+++ src/test/java/com/hivemq/extensions/cluster/discovery/s3/S3DiscoveryExtensionMainTest.java
@@ -23,7 +23,6 @@ import com.hivemq.extension.sdk.api.parameter.ExtensionStartOutput;
 import com.hivemq.extension.sdk.api.parameter.ExtensionStopInput;
 import com.hivemq.extension.sdk.api.parameter.ExtensionStopOutput;
 import com.hivemq.extensions.cluster.discovery.s3.logging.ExtensionLogging;
-import com.hivemq.extensions.cluster.discovery.s3.metrics.S3DiscoveryMetrics;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
